refactor: rename learning-content.js to replicate-learning-content.js
