+++ src/replicate-learning-content.js
